Create README.md
Put password guesses in passwords.txt then run decrypt.sh.

It will try to decrypt 1.bin with all passwords and all the ciphers in ciphers.txt. I had to remove a couple ciphers because they always return status 0, so I hope its not one of those. If it decrypts 1.bin it will say "Success!" and decrypt the rest as well.